# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /event-settings:
    get:
      operationId: getEventSettings
      tags: [settings]
      summary: Get event defaults
      responses:
        "200":
          description: Event settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSettings"
    put:
      operationId: updateEventSettings
      tags: [settings]
      summary: Save event defaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSettingsInput"
      responses:
        "200":
          description: Saved event settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSettings"
  /icloud-settings:
    get:
      operationId: getIcloudSettings
      tags: [settings]
      summary: Get saved iCloud connection metadata
      responses:
        "200":
          description: iCloud connection state; never includes the password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICloudSettings"
    put:
      operationId: saveIcloudSettings
      tags: [settings]
      summary: Save iCloud credentials and selected calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ICloudSettingsInput"
      responses:
        "200":
          description: Saved connection metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICloudSettings"
    delete:
      operationId: deleteIcloudCredentials
      tags: [settings]
      summary: Remove saved iCloud credentials
      responses:
        "200":
          description: Credentials removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICloudSettings"
  /icloud-settings/test:
    post:
      operationId: testIcloudConnection
      tags: [settings]
      summary: Test CalDAV credentials and discover calendars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestICloudConnectionInput"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICloudConnectionTest"
...
components:
  schemas:
...
    EventSettings:
      type: object
      required: [title, location, url, note, inviteAsAttendees, inviteesPerEvent, timezone, durationMinutes, minimumLeadMinutes, automaticTimeSelection, startAt]
      properties:
        title: { type: string }
        location: { type: string }
        url: { type: string }
        note: { type: string }
        inviteAsAttendees: { type: boolean }
        inviteesPerEvent: { type: integer, minimum: 1, maximum: 5000 }
        timezone: { type: string }
        durationMinutes: { type: integer }
        minimumLeadMinutes: { type: integer }
        automaticTimeSelection: { type: boolean }
        startAt: { type: ["string", "null"], format: date-time }
    EventSettingsInput:
      type: object
      required: [title, location, url, note, inviteAsAttendees, inviteesPerEvent, timezone, durationMinutes, minimumLeadMinutes, automaticTimeSelection, startAt]
      properties:
        title: { type: string, maxLength: 200 }
        location: { type: string, maxLength: 300 }
        url: { type: string, maxLength: 2000 }
        note: { type: string, maxLength: 5000 }
        inviteAsAttendees: { type: boolean }
        inviteesPerEvent: { type: integer, minimum: 1, maximum: 5000 }
        timezone: { type: string, maxLength: 100 }
        durationMinutes: { type: integer, minimum: 5, maximum: 1440 }
        minimumLeadMinutes: { type: integer, minimum: 15, maximum: 10080 }
        automaticTimeSelection: { type: boolean }
        startAt: { type: ["string", "null"], format: date-time }
    CalendarOption:
      type: object
      required: [href, name]
      properties:
        href: { type: string }
        name: { type: string }
        description: { type: string }
    ICloudSettings:
      type: object
      required: [email, calendarName, timezone, status, hasStoredPassword, calendars, invitationMode]
      properties:
        email: { type: string }
        calendarName: { type: string }
        timezone: { type: string }
        status:
          type: string
          enum: [not-connected, connected, authentication-failed, calendar-unavailable]
        hasStoredPassword: { type: boolean }
        invitationMode:
          type: string
          enum: [caldav-scheduling, calendar-only]
        calendars:
          type: array
          items: { $ref: "#/components/schemas/CalendarOption" }
    ICloudSettingsInput:
      type: object
      required: [email, calendarName, timezone]
      properties:
        email: { type: string, format: email }
        appPassword: { type: string, minLength: 1, maxLength: 100 }
        calendarName: { type: string, minLength: 1, maxLength: 200 }
        calendarHref: { type: string, minLength: 1, maxLength: 2000 }
        timezone: { type: string, minLength: 1, maxLength: 100 }
    TestICloudConnectionInput:
      type: object
      required: [email, appPassword]
      properties:
        email: { type: string, format: email }
        appPassword: { type: string, minLength: 1, maxLength: 100 }
    ICloudConnectionTest:
      type: object
      required: [status, calendars, invitationMode, message]
      properties:
        status:
          type: string
          enum: [connected, authentication-failed, calendar-unavailable, server-unavailable]
        calendars:
          type: array
          items: { $ref: "#/components/schemas/CalendarOption" }
        invitationMode:
          type: string
          enum: [caldav-scheduling, calendar-only]
        message: { type: string }
...